+++ environment.d.ts
@@ -3,5 +3,6 @@ namespace NodeJS {
     DART_API_KEY: string;
     CORP_CODE_FILENAME: string;
     CORP_CODE_PATH: string;
+    FINANCIAL_STATEMENT_PATH: string;
   }
 }
