+++ src/server/constants/apiUrls.ts
@@ -1,4 +1,3 @@
-const DART_API_KEY = process.env.DART_API_KEY;
 const DART_DOMAIN = 'https://opendart.fss.or.kr';
 
 const DART_API_URLS = {
